Moving MeasurementMetadata and Identifier out of measurement, to avoid circular dependencies. Final location not yet decided.
