+++ src/mlte/measurement/memory/local_process_memory_consumption.py
@@ -9,7 +9,7 @@ import subprocess
 from typing import Dict, Any
 
 from ..process_measurement import ProcessMeasurement
-from ..measurement_metadata import MeasurementMetadata
+from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
 from mlte.value import Value
 from mlte.validation import (
     Validator,
