+++ src/mlte/value/real.py
@@ -8,7 +8,7 @@ from typing import Dict, Any
 
 from .value import Value
 from ..validation import Validator, ValidationResult, Success, Failure
-from mlte.measurement.measurement_metadata import MeasurementMetadata
+from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
 
 
 class Real(Value):
