+++ test/measurement/test_external_measurement.py
@@ -4,8 +4,8 @@ Unit test for ExternalMeasurement.
 import pytest
 
 from mlte.value import Integer
-from mlte.measurement.measurement_metadata import MeasurementMetadata
-from mlte.measurement.identifier import Identifier
+from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
+from mlte.identifier import Identifier
 from mlte.measurement import ExternalMeasurement
