+++ test/measurement/value/test_opaque.py
@@ -6,9 +6,9 @@ import pytest
 
 import mlte
 from mlte.measurement import Measurement
-from mlte.measurement.measurement_metadata import MeasurementMetadata
-from mlte.measurement.identifier import Identifier
-from mlte.measurement.result import Opaque
+from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
+from mlte.identifier import Identifier
+from mlte.value import Opaque
 
 
 class DummyMeasurementOpaque(Measurement):
