+++ test/measurement/value/test_real.py
@@ -6,8 +6,8 @@ import pytest
 
 import mlte
 from mlte.measurement import Measurement
-from mlte.measurement.measurement_metadata import MeasurementMetadata
-from mlte.measurement.identifier import Identifier
+from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
+from mlte.identifier import Identifier
 from mlte.value import Real
